Some progress debugging Simplex.hpp
Larger examples still fail

+++ include/Math/Array.hpp
@@ -372,7 +372,7 @@ template <class T, class S> struct MutArray : Array<T, S> {
   [[gnu::flatten]] constexpr auto operator<<(const std::integral auto b)
     -> decltype(auto) {
     if constexpr (std::integral<S> || std::is_same_v<S, StridedRange>) {
-      for (size_t c = 0, L = this->sz; c < L; ++c) (*this)[c] = b;
+      for (size_t c = 0, L = size_t(this->sz); c < L; ++c) (*this)[c] = b;
     } else {
       for (size_t r = 0; r < this->numRow(); ++r)
         for (size_t c = 0; c < this->numCol(); ++c) (*this)(r, c) = b;
@@ -438,7 +438,7 @@ template <class T, class S> struct MutArray : Array<T, S> {
   [[gnu::flatten]] constexpr auto operator*=(const std::integral auto b)
     -> decltype(auto) {
     if constexpr (std::integral<S>) {
-      for (size_t c = 0, L = this->sz; c < L; ++c) (*this)[c] *= b;
+      for (size_t c = 0, L = size_t(this->sz); c < L; ++c) (*this)[c] *= b;
     } else {
       for (size_t r = 0; r < this->numRow(); ++r)
         for (size_t c = 0; c < this->numCol(); ++c) (*this)(r, c) *= b;
@@ -448,7 +448,7 @@ template <class T, class S> struct MutArray : Array<T, S> {
   [[gnu::flatten]] constexpr auto operator/=(const std::integral auto b)
     -> decltype(auto) {
     if constexpr (std::integral<S>) {
-      for (size_t c = 0, L = this->sz; c < L; ++c) (*this)[c] /= b;
+      for (size_t c = 0, L = size_t(this->sz); c < L; ++c) (*this)[c] /= b;
     } else {
       for (size_t r = 0; r < this->numRow(); ++r)
         for (size_t c = 0; c < this->numCol(); ++c) (*this)(r, c) /= b;
